Move hdf5 types into another file.
